+++ .eslintignore
@@ -1,5 +1,5 @@
 node_modules
 
 docs
-lib
+dist
 out
