+++ .prettierignore
@@ -1,5 +1,5 @@
 docs
-lib
+dist
 build
 out
 .next
