+++ packages/core/base/.gitignore
@@ -1,1 +1,1 @@
-lib
+dist
