+++ packages/core/base/.prettierignore
@@ -1,1 +1,1 @@
-lib
+dist
