+++ packages/core/base/src/types/embed/crypto.ts
