+++ packages/core/base/src/types/embed/fiat.ts
