+++ packages/core/base/src/types/embed/index.ts
