+++ packages/core/base/tsconfig.json
@@ -3,6 +3,6 @@
     "include": ["src"],
     "compilerOptions": {
         "noEmit": false,
-        "outDir": "lib/esm"
+        "outDir": "dist/esm"
     }
 }
