+++ packages/starter/vanilla-ui-starter/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
     <head>
         <title>Crossmint pay button</title>
-        <script src="https://unpkg.com/@crossmint/client-sdk-vanilla-ui@0.0.2/lib/index.global.js"></script>
+        <script src="https://unpkg.com/@crossmint/client-sdk-vanilla-ui@1.0.1-alpha.6/lib/index.global.js"></script>
     </head>
     <body>
         <h1>Crossmint pay button</h1>
