+++ packages/starter/vanilla-ui-starter/paymentElement.html
@@ -2,7 +2,7 @@
 <html lang="en">
     <head>
         <title>Crossmint Payment Element</title>
-        <script src="../../ui/vanilla-ui/lib/index.global.js"></script>
+        <script src="../../ui/vanilla-ui/dist/index.global.js"></script>
     </head>
     <body>
         <h1>Crossmint payment element</h1>
