+++ packages/ui/react-ui/.gitignore
@@ -1,4 +1,4 @@
-lib
+dist
 package
 package.tgz
 yarn-error.log
