+++ packages/ui/react-ui/.prettierignore
@@ -1,1 +1,1 @@
-lib
+dist
