+++ packages/ui/react-ui/__tests__/CrossmintPayButton.spec.tsx
@@ -2,10 +2,8 @@ import "@testing-library/jest-dom";
 import { act, fireEvent, render, screen } from "@testing-library/react";
 import React from "react";
 
-import { baseUrls } from "@crossmint/client-sdk-base";
-
-import { CrossmintPayButton } from "../src/CrossmintPayButton";
-import { LIB_VERSION } from "../src/version";
+import { CrossmintPayButton } from "../src/components/hosted";
+import { LIB_VERSION } from "../src/consts/version";
 
 // TODO(#60): create a global service for this to work everywhere and to be able to customize resolved/rejected responses
 const fetchReturns = Promise.resolve({
