+++ packages/ui/react-ui/package.json
@@ -6,17 +6,17 @@
     "repository": "https://github.com/CrossMint/crossmint-client-sdk",
     "type": "module",
     "sideEffects": false,
-    "main": "lib/index.cjs",
-    "module": "lib/index.js",
-    "types": "lib/index.d.ts",
+    "main": "./dist/index.cjs",
+    "module": "./dist/index.js",
+    "types": "./dist/index.d.ts",
     "exports": {
         ".": {
-            "import": "./lib/index.js",
-            "require": "./lib/index.cjs"
+            "import": "./dist/index.js",
+            "require": "./dist/index.cjs"
         }
     },
     "files": [
-        "lib",
+        "dist",
         "src",
         "LICENSE"
     ],
@@ -24,11 +24,11 @@
         "access": "public"
     },
     "scripts": {
-        "create-version-file": "node -p \"'export const LIB_VERSION = ' + JSON.stringify(require('./package.json').version) + ';'\" > src/version.ts",
+        "create-version-file": "node -p \"'export const LIB_VERSION = ' + JSON.stringify(require('./package.json').version) + ';'\" > src/consts/version.ts",
         "version": "yarn run create-version-file && git add .",
         "prebuild": "yarn run create-version-file",
-        "clean": "shx rm -rf lib/*",
-        "build": "yarn clean && tsup src/index.ts --external react,react-dom --format esm,cjs --outDir ./lib --minify --dts",
+        "clean": "shx rm -rf dist/*",
+        "build": "yarn clean && tsup src/index.ts --external react,react-dom --format esm,cjs --outDir ./dist --minify --dts",
         "test": "jest"
     },
     "dependencies": {
@@ -46,7 +46,7 @@
         "@types/react": "17.0.34",
         "@types/react-dom": "^17.0.11",
         "jest": "^27.5.1",
-        "node-fetch": "^3.2.0",
+        "node-fetch": "3.2.10",
         "ts-jest": "^27.1.3",
         "ts-node-dev": "^1.1.8"
     }
