+++ packages/ui/react-ui/src/components/hosted/index.ts
@@ -1,0 +1,1 @@
+export * from "./CrossmintPayButton";
