+++ packages/ui/react-ui/src/components/index.ts
@@ -1,0 +1,5 @@
+export * from "./CrossmintNFTCollectionView";
+export * from "./CrossmintNFTDetail";
+
+export * from "./embed";
+export * from "./hosted";
