+++ packages/ui/react-ui/src/hooks/useEnvironment.ts
@@ -1,0 +1,10 @@
+import { useEffect, useState } from "react";
+
+export default function useEnvironment() {
+    const [isServerSideRendering, setIsServerSideRendering] = useState(true);
+    useEffect(() => {
+        setIsServerSideRendering(false);
+    }, []);
+
+    return { isServerSideRendering };
+}
