+++ packages/ui/react-ui/src/index.ts
@@ -1,7 +1,4 @@
-export * from "./CrossmintPayButton";
-export * from "./CrossmintNFTCollectionView";
-export * from "./CrossmintNFTDetail";
-export * from "./EmbeddedCheckout";
+export * from "./components";
 
 export { CheckoutEvents, useCrossmintEvents } from "@crossmint/client-sdk-base";
 export type { CrossmintCheckoutEvent, CheckoutEventMap, CrossmintCheckoutEventUnion } from "@crossmint/client-sdk-base";
