--- packages/ui/react-ui/src/types.ts
@@ -1,8 +1,0 @@
-import { CSSProperties, MouseEvent } from "react";
-
-import { CrossmintPayButtonProps } from "@crossmint/client-sdk-base";
-
-export type CrossmintPayButtonReactProps = CrossmintPayButtonProps & {
-    onClick?: (e: MouseEvent<HTMLButtonElement>) => void;
-    style?: CSSProperties;
-};
