--- packages/ui/react-ui/src/useEnvironment.ts
@@ -1,10 +1,0 @@
-import { useEffect, useState } from "react";
-
-export default function useEnvironment() {
-    const [isServerSideRendering, setIsServerSideRendering] = useState(true);
-    useEffect(() => {
-        setIsServerSideRendering(false);
-    }, []);
-
-    return { isServerSideRendering };
-}
