+++ packages/ui/vanilla-ui/.gitignore
@@ -1,1 +1,1 @@
-lib
+dist
