+++ packages/ui/vanilla-ui/package.json
@@ -8,17 +8,17 @@
     "repository": "https://github.com/Crossmint/crossmint-client-sdk",
     "type": "module",
     "sideEffects": false,
-    "main": "lib/index.cjs",
-    "module": "lib/index.js",
-    "types": "lib/index.d.ts",
+    "main": "./dist/index.cjs",
+    "module": "./dist/index.js",
+    "types": "./dist/index.d.ts",
     "exports": {
         ".": {
-            "import": "./lib/index.js",
-            "require": "./lib/index.cjs"
+            "import": "./dist/index.js",
+            "require": "./dist/index.cjs"
         }
     },
     "files": [
-        "lib",
+        "dist",
         "src",
         "LICENSE"
     ],
@@ -26,11 +26,11 @@
         "access": "public"
     },
     "scripts": {
-        "create-version-file": "node -p \"'export const LIB_VERSION = ' + JSON.stringify(require('./package.json').version) + ';'\" > src/version.ts",
+        "create-version-file": "node -p \"'export const LIB_VERSION = ' + JSON.stringify(require('./package.json').version) + ';'\" > src/consts/version.ts",
         "version": "yarn run create-version-file && git add .",
         "prebuild": "yarn run create-version-file",
-        "clean": "shx rm -rf lib/*",
-        "build": "yarn clean && tsup src/index.ts --format esm,cjs,iife --outDir ./lib --minify --dts"
+        "clean": "shx rm -rf dist/*",
+        "build": "yarn clean && tsup src/index.ts --format esm,cjs,iife --outDir ./dist --minify --dts"
     },
     "bugs": {
         "url": "https://github.com/Crossmint/crossmint-client-sdk/issues"
