+++ packages/ui/vanilla-ui/src/components/embed/index.ts
@@ -1,0 +1,1 @@
+export * from "./CrossmintPaymentElement";
