+++ packages/ui/vanilla-ui/src/components/index.ts
@@ -1,0 +1,2 @@
+export * from "./embed";
+export * from "./hosted";
