+++ packages/ui/vanilla-ui/src/index.ts
@@ -1,4 +1,4 @@
-export * from "./CrossmintPayButton";
-export * from "./CrossmintPaymentElement";
+export * from "./components";
+
 export { CheckoutEvents, useCrossmintEvents } from "@crossmint/client-sdk-base";
 export type { CrossmintCheckoutEvent, CheckoutEventMap, CrossmintCheckoutEventUnion } from "@crossmint/client-sdk-base";
