--- packages/ui/vanilla-ui/src/types.ts
@@ -1,6 +1,0 @@
-import { CrossmintPayButtonProps } from "@crossmint/client-sdk-base";
-
-export type CrossmintPayButtonLitProps = CrossmintPayButtonProps & {
-    onClick?: (e: any) => void;
-};
-
