+++ packages/ui/vue-ui/.prettierignore
@@ -1,2 +1,2 @@
-lib
+dist
 tsconfig.tsbuildinfo
