+++ packages/ui/vue-ui/vite.config.ts
@@ -6,6 +6,7 @@ import dts from "vite-plugin-dts";
 export default defineConfig({
     plugins: [vue(), dts()],
     build: {
+        outDir: "dist",
         lib: {
             entry: resolve(__dirname, "./src/index.ts"),
             name: "CrossmintVueSDKUi",
